+++ src/utils/betterArrays/dicts/int_dict.c
@@ -115,7 +115,7 @@ bool remove_from_int_dictionary(IntDictionary* dict, int key) {
 }
 
 void iterate_int_dictionary(IntDictionary* dict, void (*callback)(int key, void* value)) {
-    for (size_t i = 0; i < dict->capacity; i++) {
+    for (int i = 0; i < dict->capacity; i++) {
         Node* current = dict->array[i];
         while (current) {
             callback(current->key, current->value);
