+++ src/utils/betterArrays/queue.h
@@ -14,10 +14,10 @@ typedef struct queue_s Queue;
 // Function declarations
 Queue* create_queue();
 bool queue_is_empty(Queue *q);
-bool queue_enqueue(Queue *q, void *data, size_t data_size);
-bool queue_dequeue(Queue *q, void *out_data, size_t data_size);
-bool queue_peek(Queue *q, void *out_data, size_t data_size);
-size_t queue_size(Queue *q);
+bool queue_enqueue(Queue *q, void *data, int data_size);
+bool queue_dequeue(Queue *q, void *out_data, int data_size);
+bool queue_peek(Queue *q, void *out_data, int data_size);
+int queue_size(Queue *q);
 void queue_clear(Queue *q);
 void queue_destroy(Queue *q);
